Move Symbol to interner.rs
