+++ mimium-lang/src/interner.rs
@@ -1,4 +1,9 @@
-use std::{cell::RefCell, collections::BTreeMap, hash::Hash};
+use std::{
+    cell::RefCell,
+    collections::BTreeMap,
+    fmt::{self, Display},
+    hash::Hash,
+};
 
 use slotmap::{Key, KeyData, SlotMap};
 use string_interner::{backend::StringBackend, StringInterner};
@@ -64,6 +69,44 @@ where
     F: FnOnce(&mut SessionGlobals) -> R,
 {
     SESSION_GLOBALS.with_borrow_mut(f)
+}
+
+#[derive(Default, Copy, Clone, PartialEq, Debug, Hash, Eq)]
+pub struct Symbol(pub usize); //Symbol Trait is implemented on usize
+
+pub trait ToSymbol {
+    fn to_symbol(&self) -> Symbol;
+}
+
+impl<T: AsRef<str>> ToSymbol for T {
+    fn to_symbol(&self) -> Symbol {
+        Symbol(with_session_globals(|session_globals| {
+            session_globals.symbol_interner.get_or_intern(self.as_ref())
+        }))
+    }
+}
+
+impl Symbol {
+    pub fn as_str(&self) -> &str {
+        with_session_globals(|session_globals| unsafe {
+            // This transmute is needed to convince the borrow checker. Since
+            // the session_global should exist until the end of the session,
+            // this &str should live sufficiently long.
+            std::mem::transmute::<&str, &str>(
+                session_globals
+                    .symbol_interner
+                    .resolve(self.0)
+                    .expect("invalid symbol"),
+            )
+        })
+    }
+}
+
+// Note: to_string() is auto-implemented by this
+impl Display for Symbol {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.as_str())
+    }
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
