+++ mimium-lang/src/runtime.rs
@@ -1,5 +1,5 @@
-use crate::ast::ToSymbol;
 use crate::compiler;
+use crate::interner::ToSymbol;
 use crate::utils::{error::ReportableError, metadata::Span};
 
 pub mod scheduler;
