+++ mimium-lang/src/runtime/vm.rs
@@ -19,7 +19,7 @@ use program::OpenUpValue;
 pub use program::{FuncProto, Program};
 
 use crate::{
-    ast::{Symbol, ToSymbol},
+    interner::{Symbol, ToSymbol},
     types::TypeSize,
 };
 pub type RawVal = u64;
